==> 0_Realestate_API ACT.ipynb <==
# The aim of this script is to download all rental property data and all available property from realestate website.
# the looping method doesn't work, hence the long winded version

# pip install -e git+https://github.com/tomquirk/realestate-com-au-api.git#egg=realestate_com_au_api

from realestate_com_au import RealestateComAu
api = RealestateComAu()

# Get property listings
# listings = api.search(locations=["seventeen seventy, qld 4677"], channel="buy", keywords=["tenant"], exclude_keywords=["pool"])

# https://www.matthewproctor.com/australian_postcodes
# df_postcode = pd.read_excel (r'/Users/joezhou/Downloads/Realestate/australian_postcodes.xlsx', sheet_name='Sheet1')

# sample test suburb
# df_postcode_subset = df_postcode.loc[(df_postcode['Locality']=='LIDCOMBE') & (df_postcode['Postcode']==2141) ]

# df_postcode_subset["VAR_Find"] = '["'+df_postcode_subset['Postcode'].apply(str)+","+df_postcode_subset['Locality']+'"]'
# list_postcode = list(df_postcode_subset['VAR_Find'].dropna())

import pandas as pd

# try a loop approach by suburb
#note: code below doesn't work for some reason, it returns an arbitary list of values

DF_RENT = pd.DataFrame()
DF_ALL = pd.DataFrame()

def DWN_INF(Find_Text):    
    global DF_RENT
    try:
        df_imp_rent = api.search(locations=Find_Text, channel="rent")
        df_imp_rent = pd.DataFrame(df_imp_rent)
        
        from datetime import datetime as dt
        dte = dt.now().strftime('%Y%m%d')
        fname = "/Users/joezhou/Downloads/loop_test{}.xlsx".format(dte)
        df_imp_rent.to_excel(fname)

        df_imp_all = api.search(locations=Find_Text)
        df_imp_all = pd.DataFrame(df_imp_all)

        DF_RENT = DF_RENT.append(df_imp_rent, ignore_index=True) 
        DF_ALL = DF_ALL.append(df_imp_all, ignore_index=True) 

        print("found for:",Find_Text)

    except Exception:
        print("Error with:",Find_Text)
        pass    
    return()

# Download all
# for Find_Var in list_postcode:
#    DWN_INF(Find_Var)


# ACT - rental
DF_RENT = pd.DataFrame()

df_out = api.search(locations=['2540,BAMARANG'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2540-BAMARANG")
df_out = api.search(locations=['2600,BARTON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2600-BARTON")
df_out = api.search(locations=['2601,ACTON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2601-ACTON")
df_out = api.search(locations=['2602,AINSLIE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2602-AINSLIE")
df_out = api.search(locations=['2603,FORREST'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2603-FORREST")
df_out = api.search(locations=['2604,CAUSEWAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2604-CAUSEWAY")
df_out = api.search(locations=['2605,CURTIN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2605-CURTIN")
df_out = api.search(locations=['2606,CHIFLEY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2606-CHIFLEY")
df_out = api.search(locations=['2607,FARRER'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2607-FARRER")
df_out = api.search(locations=['2609,CANBERRA AIRPORT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2609-CANBERRA AIRPORT")
df_out = api.search(locations=['2611,BIMBERI'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2611-BIMBERI")
df_out = api.search(locations=['2612,BRADDON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2612-BRADDON")
df_out = api.search(locations=['2614,ARANDA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2614-ARANDA")
df_out = api.search(locations=['2615,CHARNWOOD'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2615-CHARNWOOD")
df_out = api.search(locations=['2617,BELCONNEN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2617-BELCONNEN")
df_out = api.search(locations=['2618,HALL'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2618-HALL")
df_out = api.search(locations=['2620,BEARD'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2620-BEARD")
df_out = api.search(locations=['2900,GREENWAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2900-GREENWAY")
df_out = api.search(locations=['2902,KAMBAH'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2902-KAMBAH")
df_out = api.search(locations=['2903,ERINDALE CENTRE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2903-ERINDALE CENTRE")
df_out = api.search(locations=['2904,FADDEN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2904-FADDEN")
df_out = api.search(locations=['2905,BONYTHON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2905-BONYTHON")
df_out = api.search(locations=['2906,BANKS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2906-BANKS")
df_out = api.search(locations=['2911,CRACE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2911-CRACE")
df_out = api.search(locations=['2912,GUNGAHLIN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2912-GUNGAHLIN")
df_out = api.search(locations=['2913,CASEY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2913-CASEY")
df_out = api.search(locations=['2914,AMAROO'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:2914-AMAROO")


DF_RENT1 = DF_RENT.reset_index()
DF_RENT2 = DF_RENT1.drop(columns=['id','index','images','images_floorplans','listers'])
DF_RENT_UNIQ = DF_RENT2.drop_duplicates()

DF_RENT_UNIQ.to_pickle("/Users/joezhou/Downloads/Realestate/R_RENT_ACT.pkl")

# ACT - all
DF_ALL = pd.DataFrame()

df_out = api.search(locations=['2540,BAMARANG'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2540-BAMARANG")
df_out = api.search(locations=['2600,BARTON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2600-BARTON")
df_out = api.search(locations=['2601,ACTON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2601-ACTON")
df_out = api.search(locations=['2602,AINSLIE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2602-AINSLIE")
df_out = api.search(locations=['2603,FORREST'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2603-FORREST")
df_out = api.search(locations=['2604,CAUSEWAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2604-CAUSEWAY")
df_out = api.search(locations=['2605,CURTIN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2605-CURTIN")
df_out = api.search(locations=['2606,CHIFLEY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2606-CHIFLEY")
df_out = api.search(locations=['2607,FARRER'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2607-FARRER")
df_out = api.search(locations=['2609,CANBERRA AIRPORT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2609-CANBERRA AIRPORT")
df_out = api.search(locations=['2611,BIMBERI'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2611-BIMBERI")
df_out = api.search(locations=['2612,BRADDON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2612-BRADDON")
df_out = api.search(locations=['2614,ARANDA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2614-ARANDA")
df_out = api.search(locations=['2615,CHARNWOOD'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2615-CHARNWOOD")
df_out = api.search(locations=['2617,BELCONNEN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2617-BELCONNEN")
df_out = api.search(locations=['2618,HALL'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2618-HALL")
df_out = api.search(locations=['2620,BEARD'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2620-BEARD")
df_out = api.search(locations=['2900,GREENWAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2900-GREENWAY")
df_out = api.search(locations=['2902,KAMBAH'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2902-KAMBAH")
df_out = api.search(locations=['2903,ERINDALE CENTRE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2903-ERINDALE CENTRE")
df_out = api.search(locations=['2904,FADDEN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2904-FADDEN")
df_out = api.search(locations=['2905,BONYTHON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2905-BONYTHON")
df_out = api.search(locations=['2906,BANKS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2906-BANKS")
df_out = api.search(locations=['2911,CRACE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2911-CRACE")
df_out = api.search(locations=['2912,GUNGAHLIN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2912-GUNGAHLIN")
df_out = api.search(locations=['2913,CASEY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2913-CASEY")
df_out = api.search(locations=['2914,AMAROO'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:2914-AMAROO")


DF_ALL1 = DF_ALL.reset_index()
DF_ALL2 = DF_ALL1.drop(columns=['id','index','images','images_floorplans','listers'])
DF_ALL_UNIQ = DF_ALL2.drop_duplicates()

DF_ALL.to_pickle("/Users/joezhou/Downloads/Realestate/R_ALL_ACT.pkl")
