==> 0_Realestate_API NT.ipynb <==
# The aim of this script is to download all rental property data and all available property from realestate website.
# the looping method doesn't work, hence the long winded version

# pip install -e git+https://github.com/tomquirk/realestate-com-au-api.git#egg=realestate_com_au_api

from realestate_com_au import RealestateComAu
api = RealestateComAu()

# Get property listings
# listings = api.search(locations=["seventeen seventy, qld 4677"], channel="buy", keywords=["tenant"], exclude_keywords=["pool"])

# https://www.matthewproctor.com/australian_postcodes
# df_postcode = pd.read_excel (r'/Users/joezhou/Downloads/Realestate/australian_postcodes.xlsx', sheet_name='Sheet1')

# sample test suburb
# df_postcode_subset = df_postcode.loc[(df_postcode['Locality']=='LIDCOMBE') & (df_postcode['Postcode']==2141) ]

# df_postcode_subset["VAR_Find"] = '["'+df_postcode_subset['Postcode'].apply(str)+","+df_postcode_subset['Locality']+'"]'
# list_postcode = list(df_postcode_subset['VAR_Find'].dropna())

import pandas as pd

# try a loop approach by suburb
#note: code below doesn't work for some reason, it returns an arbitary list of values

DF_RENT = pd.DataFrame()
DF_ALL = pd.DataFrame()

def DWN_INF(Find_Text):    
    global DF_RENT
    try:
        df_imp_rent = api.search(locations=Find_Text, channel="rent")
        df_imp_rent = pd.DataFrame(df_imp_rent)
        
        from datetime import datetime as dt
        dte = dt.now().strftime('%Y%m%d')
        fname = "/Users/joezhou/Downloads/loop_test{}.xlsx".format(dte)
        df_imp_rent.to_excel(fname)

        df_imp_all = api.search(locations=Find_Text)
        df_imp_all = pd.DataFrame(df_imp_all)

        DF_RENT = DF_RENT.append(df_imp_rent, ignore_index=True) 
        DF_ALL = DF_ALL.append(df_imp_all, ignore_index=True) 

        print("found for:",Find_Text)

    except Exception:
        print("Error with:",Find_Text)
        pass    
    return()

# Download all
# for Find_Var in list_postcode:
#    DWN_INF(Find_Var)


#  NT - rental
DF_RENT = pd.DataFrame()

df_out = api.search(locations=['0810,ALAWA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0810-ALAWA")
df_out = api.search(locations=['0812,ANULA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0812-ANULA")
df_out = api.search(locations=['0820,BAGOT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0820-BAGOT")
df_out = api.search(locations=['0822,ACACIA HILLS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0822-ACACIA HILLS")
df_out = api.search(locations=['0829,HOLTZE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0829-HOLTZE")
df_out = api.search(locations=['0830,ARCHER'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0830-ARCHER")
df_out = api.search(locations=['0832,BAKEWELL'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0832-BAKEWELL")
df_out = api.search(locations=['0834,VIRGINIA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0834-VIRGINIA")
df_out = api.search(locations=['0836,HERBERT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0836-HERBERT")
df_out = api.search(locations=['0837,MANTON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0837-MANTON")
df_out = api.search(locations=['0839,COOLALINGA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0839-COOLALINGA")
df_out = api.search(locations=['0850,COSSACK'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0850-COSSACK")
df_out = api.search(locations=['0852,ARNOLD'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0852-ARNOLD")
df_out = api.search(locations=['0853,TINDAL'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0853-TINDAL")
df_out = api.search(locations=['0854,BORROLOOLA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0854-BORROLOOLA")
df_out = api.search(locations=['0862,AVON DOWNS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0862-AVON DOWNS")
df_out = api.search(locations=['0870,ALICE SPRINGS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0870-ALICE SPRINGS")
df_out = api.search(locations=['0872,AHERRENGE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0872-AHERRENGE")
df_out = api.search(locations=['0873,AMOONGUNA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0873-AMOONGUNA")
df_out = api.search(locations=['0874,IRLPME'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0874-IRLPME")
df_out = api.search(locations=['0875,FLYNN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0875-FLYNN")
df_out = api.search(locations=['0880,GAPUWIYAK'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0880-GAPUWIYAK")
df_out = api.search(locations=['0885,ALYANGULA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:0885-ALYANGULA")
df_out = api.search(locations=['4825,ALEXANDRIA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:4825-ALEXANDRIA")

DF_RENT1 = DF_RENT.reset_index()
DF_RENT2 = DF_RENT1.drop(columns=['id','index','images','images_floorplans','listers'])
DF_RENT_UNIQ = DF_RENT2.drop_duplicates()

DF_RENT_UNIQ.to_pickle("/Users/joezhou/Downloads/Realestate/R_RENT_NT.pkl")

# NT - all
DF_ALL = pd.DataFrame()

df_out = api.search(locations=['0810,ALAWA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0810-ALAWA")
df_out = api.search(locations=['0812,ANULA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0812-ANULA")
df_out = api.search(locations=['0820,BAGOT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0820-BAGOT")
df_out = api.search(locations=['0822,ACACIA HILLS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0822-ACACIA HILLS")
df_out = api.search(locations=['0829,HOLTZE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0829-HOLTZE")
df_out = api.search(locations=['0830,ARCHER'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0830-ARCHER")
df_out = api.search(locations=['0832,BAKEWELL'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0832-BAKEWELL")
df_out = api.search(locations=['0834,VIRGINIA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0834-VIRGINIA")
df_out = api.search(locations=['0836,HERBERT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0836-HERBERT")
df_out = api.search(locations=['0837,MANTON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0837-MANTON")
df_out = api.search(locations=['0839,COOLALINGA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0839-COOLALINGA")
df_out = api.search(locations=['0850,COSSACK'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0850-COSSACK")
df_out = api.search(locations=['0852,ARNOLD'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0852-ARNOLD")
df_out = api.search(locations=['0853,TINDAL'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0853-TINDAL")
df_out = api.search(locations=['0854,BORROLOOLA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0854-BORROLOOLA")
df_out = api.search(locations=['0862,AVON DOWNS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0862-AVON DOWNS")
df_out = api.search(locations=['0870,ALICE SPRINGS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0870-ALICE SPRINGS")
df_out = api.search(locations=['0872,AHERRENGE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0872-AHERRENGE")
df_out = api.search(locations=['0873,AMOONGUNA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0873-AMOONGUNA")
df_out = api.search(locations=['0874,IRLPME'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0874-IRLPME")
df_out = api.search(locations=['0875,FLYNN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0875-FLYNN")
df_out = api.search(locations=['0880,GAPUWIYAK'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0880-GAPUWIYAK")
df_out = api.search(locations=['0885,ALYANGULA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:0885-ALYANGULA")
df_out = api.search(locations=['4825,ALEXANDRIA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:4825-ALEXANDRIA")

DF_ALL1 = DF_ALL.reset_index()
DF_ALL2 = DF_ALL1.drop(columns=['id','index','images','images_floorplans','listers'])
DF_ALL_UNIQ = DF_ALL2.drop_duplicates()

DF_ALL.to_pickle("/Users/joezhou/Downloads/Realestate/R_ALL_NT.pkl")
