==> 0_Realestate_API TAS.ipynb <==
# The aim of this script is to download all rental property data and all available property from realestate website.
# the looping method doesn't work, hence the long winded version

# pip install -e git+https://github.com/tomquirk/realestate-com-au-api.git#egg=realestate_com_au_api

from realestate_com_au import RealestateComAu
api = RealestateComAu()

# Get property listings
# listings = api.search(locations=["seventeen seventy, qld 4677"], channel="buy", keywords=["tenant"], exclude_keywords=["pool"])

# https://www.matthewproctor.com/australian_postcodes
# df_postcode = pd.read_excel (r'/Users/joezhou/Downloads/Realestate/australian_postcodes.xlsx', sheet_name='Sheet1')

# sample test suburb
# df_postcode_subset = df_postcode.loc[(df_postcode['Locality']=='LIDCOMBE') & (df_postcode['Postcode']==2141) ]

# df_postcode_subset["VAR_Find"] = '["'+df_postcode_subset['Postcode'].apply(str)+","+df_postcode_subset['Locality']+'"]'
# list_postcode = list(df_postcode_subset['VAR_Find'].dropna())

import pandas as pd

# try a loop approach by suburb
#note: code below doesn't work for some reason, it returns an arbitary list of values

DF_RENT = pd.DataFrame()
DF_ALL = pd.DataFrame()

def DWN_INF(Find_Text):    
    global DF_RENT
    try:
        df_imp_rent = api.search(locations=Find_Text, channel="rent")
        df_imp_rent = pd.DataFrame(df_imp_rent)
        
        from datetime import datetime as dt
        dte = dt.now().strftime('%Y%m%d')
        fname = "/Users/joezhou/Downloads/loop_test{}.xlsx".format(dte)
        df_imp_rent.to_excel(fname)

        df_imp_all = api.search(locations=Find_Text)
        df_imp_all = pd.DataFrame(df_imp_all)

        DF_RENT = DF_RENT.append(df_imp_rent, ignore_index=True) 
        DF_ALL = DF_ALL.append(df_imp_all, ignore_index=True) 

        print("found for:",Find_Text)

    except Exception:
        print("Error with:",Find_Text)
        pass    
    return()

# Download all
# for Find_Var in list_postcode:
#    DWN_INF(Find_Var)


# TAS - rental
DF_RENT = pd.DataFrame()

df_out = api.search(locations=['7000,BATHURST STREET PO'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7000-BATHURST STREET PO")
df_out = api.search(locations=['7001,HOBART'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7001-HOBART")
df_out = api.search(locations=['7004,BATTERY POINT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7004-BATTERY POINT")
df_out = api.search(locations=['7005,DYNNYRNE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7005-DYNNYRNE")
df_out = api.search(locations=['7007,MOUNT NELSON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7007-MOUNT NELSON")
df_out = api.search(locations=['7008,CORNELIAN BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7008-CORNELIAN BAY")
df_out = api.search(locations=['7009,DERWENT PARK'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7009-DERWENT PARK")
df_out = api.search(locations=['7010,DOWSING POINT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7010-DOWSING POINT")
df_out = api.search(locations=['7011,AUSTINS FERRY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7011-AUSTINS FERRY")
df_out = api.search(locations=['7012,COLLINSVALE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7012-COLLINSVALE")
df_out = api.search(locations=['7015,FLAGSTAFF GULLY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7015-FLAGSTAFF GULLY")
df_out = api.search(locations=['7016,RISDON VALE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7016-RISDON VALE")
df_out = api.search(locations=['7017,GRASSTREE HILL'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7017-GRASSTREE HILL")
df_out = api.search(locations=['7018,BELLERIVE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7018-BELLERIVE")
df_out = api.search(locations=['7019,CLARENDON VALE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7019-CLARENDON VALE")
df_out = api.search(locations=['7020,CLIFTON BEACH'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7020-CLIFTON BEACH")
df_out = api.search(locations=['7021,LAUDERDALE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7021-LAUDERDALE")
df_out = api.search(locations=['7022,SOUTH ARM'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7022-SOUTH ARM")
df_out = api.search(locations=['7023,OPOSSUM BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7023-OPOSSUM BAY")
df_out = api.search(locations=['7024,CREMORNE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7024-CREMORNE")
df_out = api.search(locations=['7025,DULCOT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7025-DULCOT")
df_out = api.search(locations=['7026,CAMPANIA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7026-CAMPANIA")
df_out = api.search(locations=['7027,COLEBROOK'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7027-COLEBROOK")
df_out = api.search(locations=['7030,APSLEY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7030-APSLEY")
df_out = api.search(locations=['7050,ALBION HEIGHTS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7050-ALBION HEIGHTS")
df_out = api.search(locations=['7052,BLACKMANS BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7052-BLACKMANS BAY")
df_out = api.search(locations=['7053,BONNET HILL'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7053-BONNET HILL")
df_out = api.search(locations=['7054,BARRETTA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7054-BARRETTA")
df_out = api.search(locations=['7055,HUNTINGFIELD'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7055-HUNTINGFIELD")
df_out = api.search(locations=['7109,CATAMARAN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7109-CATAMARAN")
df_out = api.search(locations=['7112,ABELS BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7112-ABELS BAY")
df_out = api.search(locations=['7113,FRANKLIN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7113-FRANKLIN")
df_out = api.search(locations=['7116,BROOKS BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7116-BROOKS BAY")
df_out = api.search(locations=['7117,DOVER'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7117-DOVER")
df_out = api.search(locations=['7119,STONOR'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7119-STONOR")
df_out = api.search(locations=['7120,ANDOVER'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7120-ANDOVER")
df_out = api.search(locations=['7140,BLACK HILLS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7140-BLACK HILLS")
df_out = api.search(locations=['7150,ADVENTURE BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7150-ADVENTURE BAY")
df_out = api.search(locations=['7151,CASEY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7151-CASEY")
df_out = api.search(locations=['7155,KETTERING'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7155-KETTERING")
df_out = api.search(locations=['7162,BIRCHS BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7162-BIRCHS BAY")
df_out = api.search(locations=['7163,FLOWERPOT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7163-FLOWERPOT")
df_out = api.search(locations=['7170,ACTON PARK'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7170-ACTON PARK")
df_out = api.search(locations=['7171,MIDWAY POINT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7171-MIDWAY POINT")
df_out = api.search(locations=['7172,NUGENT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7172-NUGENT")
df_out = api.search(locations=['7173,CARLTON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7173-CARLTON")
df_out = api.search(locations=['7175,BREAM CREEK'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7175-BREAM CREEK")
df_out = api.search(locations=['7176,KELLEVIE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7176-KELLEVIE")
df_out = api.search(locations=['7177,BOOMER BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7177-BOOMER BAY")
df_out = api.search(locations=['7178,MURDUNNA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7178-MURDUNNA")
df_out = api.search(locations=['7179,EAGLEHAWK NECK'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7179-EAGLEHAWK NECK")
df_out = api.search(locations=['7180,TARANNA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7180-TARANNA")
df_out = api.search(locations=['7182,CAPE PILLAR'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7182-CAPE PILLAR")
df_out = api.search(locations=['7184,CAPE RAOUL'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7184-CAPE RAOUL")
df_out = api.search(locations=['7185,PREMAYDENA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7185-PREMAYDENA")
df_out = api.search(locations=['7186,SALTWATER RIVER'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7186-SALTWATER RIVER")
df_out = api.search(locations=['7187,KOONYA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7187-KOONYA")
df_out = api.search(locations=['7190,APSLAWN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7190-APSLAWN")
df_out = api.search(locations=['7209,ROSS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7209-ROSS")
df_out = api.search(locations=['7210,CAMPBELL TOWN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7210-CAMPBELL TOWN")
df_out = api.search(locations=['7211,CLEVELAND'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7211-CLEVELAND")
df_out = api.search(locations=['7212,BEN LOMOND'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7212-BEN LOMOND")
df_out = api.search(locations=['7213,AVOCA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7213-AVOCA")
df_out = api.search(locations=['7214,FINGAL'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7214-FINGAL")
df_out = api.search(locations=['7215,BEAUMARIS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7215-BEAUMARIS")
df_out = api.search(locations=['7216,AKAROA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7216-AKAROA")
df_out = api.search(locations=['7248,ALANVALE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7248-ALANVALE")
df_out = api.search(locations=['7249,GLEN DHU'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7249-GLEN DHU")
df_out = api.search(locations=['7250,BLACKSTONE HEIGHTS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7250-BLACKSTONE HEIGHTS")
df_out = api.search(locations=['7252,BEECHFORD'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7252-BEECHFORD")
df_out = api.search(locations=['7253,BELL BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7253-BELL BAY")
df_out = api.search(locations=['7254,BELLINGHAM'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7254-BELLINGHAM")
df_out = api.search(locations=['7255,BLUE ROCKS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7255-BLUE ROCKS")
df_out = api.search(locations=['7256,BUNGAREE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7256-BUNGAREE")
df_out = api.search(locations=['7257,CAPE BARREN ISLAND'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7257-CAPE BARREN ISLAND")
df_out = api.search(locations=['7258,BREADALBANE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7258-BREADALBANE")
df_out = api.search(locations=['7259,MYRTLE BANK'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7259-MYRTLE BANK")
df_out = api.search(locations=['7260,BLUMONT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7260-BLUMONT")
df_out = api.search(locations=['7261,BRANXHOLM'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7261-BRANXHOLM")
df_out = api.search(locations=['7262,BRIDPORT'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7262-BRIDPORT")
df_out = api.search(locations=['7263,ALBERTON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7263-ALBERTON")
df_out = api.search(locations=['7264,ANSONS BAY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7264-ANSONS BAY")
df_out = api.search(locations=['7265,BANCA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7265-BANCA")
df_out = api.search(locations=['7267,BANGOR'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7267-BANGOR")
df_out = api.search(locations=['7268,LILYDALE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7268-LILYDALE")
df_out = api.search(locations=['7270,BADGER HEAD'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7270-BADGER HEAD")
df_out = api.search(locations=['7275,BLACKWALL'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7275-BLACKWALL")
df_out = api.search(locations=['7276,GRAVELLY BEACH'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7276-GRAVELLY BEACH")
df_out = api.search(locations=['7277,BRIDGENORTH'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7277-BRIDGENORTH")
df_out = api.search(locations=['7290,HADSPEN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7290-HADSPEN")
df_out = api.search(locations=['7291,CARRICK'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7291-CARRICK")
df_out = api.search(locations=['7292,HAGLEY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7292-HAGLEY")
df_out = api.search(locations=['7300,DEVON HILLS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7300-DEVON HILLS")
df_out = api.search(locations=['7301,BISHOPSBOURNE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7301-BISHOPSBOURNE")
df_out = api.search(locations=['7302,BRACKNELL'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7302-BRACKNELL")
df_out = api.search(locations=['7303,BIRRALEE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7303-BIRRALEE")
df_out = api.search(locations=['7304,BRANDUM'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7304-BRANDUM")
df_out = api.search(locations=['7305,MERSEYLEA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7305-MERSEYLEA")
df_out = api.search(locations=['7306,ACACIA HILLS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7306-ACACIA HILLS")
df_out = api.search(locations=['7307,BAKERS BEACH'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7307-BAKERS BEACH")
df_out = api.search(locations=['7310,ABERDEEN'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7310-ABERDEEN")
df_out = api.search(locations=['7315,ABBOTSHAM'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7315-ABBOTSHAM")
df_out = api.search(locations=['7316,CAMENA'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7316-CAMENA")
df_out = api.search(locations=['7320,ACTON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7320-ACTON")
df_out = api.search(locations=['7321,BLACK RIVER'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7321-BLACK RIVER")
df_out = api.search(locations=['7322,SOMERSET'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7322-SOMERSET")
df_out = api.search(locations=['7325,CALDER'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7325-CALDER")
df_out = api.search(locations=['7330,ALCOMIE'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7330-ALCOMIE")
df_out = api.search(locations=['7331,STANLEY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7331-STANLEY")
df_out = api.search(locations=['7466,GORMANSTON'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7466-GORMANSTON")
df_out = api.search(locations=['7467,LAKE MARGARET'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7467-LAKE MARGARET")
df_out = api.search(locations=['7468,MACQUARIE HEADS'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7468-MACQUARIE HEADS")
df_out = api.search(locations=['7469,GRANVILLE HARBOUR'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7469-GRANVILLE HARBOUR")
df_out = api.search(locations=['7470,ROSEBERY'],channel="rent")
df_out = pd.DataFrame(df_out)
DF_RENT = DF_RENT.append(df_out, ignore_index=True) 
print("finished rent:7470-ROSEBERY")

DF_RENT1 = DF_RENT.reset_index()
DF_RENT2 = DF_RENT1.drop(columns=['id','index','images','images_floorplans','listers'])
DF_RENT_UNIQ = DF_RENT2.drop_duplicates()

DF_RENT_UNIQ.to_pickle("/Users/joezhou/Downloads/Realestate/R_RENT_TAS.pkl")

# TAS - all
DF_ALL = pd.DataFrame()

df_out = api.search(locations=['7000,BATHURST STREET PO'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7000-BATHURST STREET PO")
df_out = api.search(locations=['7001,HOBART'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7001-HOBART")
df_out = api.search(locations=['7004,BATTERY POINT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7004-BATTERY POINT")
df_out = api.search(locations=['7005,DYNNYRNE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7005-DYNNYRNE")
df_out = api.search(locations=['7007,MOUNT NELSON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7007-MOUNT NELSON")
df_out = api.search(locations=['7008,CORNELIAN BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7008-CORNELIAN BAY")
df_out = api.search(locations=['7009,DERWENT PARK'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7009-DERWENT PARK")
df_out = api.search(locations=['7010,DOWSING POINT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7010-DOWSING POINT")
df_out = api.search(locations=['7011,AUSTINS FERRY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7011-AUSTINS FERRY")
df_out = api.search(locations=['7012,COLLINSVALE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7012-COLLINSVALE")
df_out = api.search(locations=['7015,FLAGSTAFF GULLY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7015-FLAGSTAFF GULLY")
df_out = api.search(locations=['7016,RISDON VALE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7016-RISDON VALE")
df_out = api.search(locations=['7017,GRASSTREE HILL'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7017-GRASSTREE HILL")
df_out = api.search(locations=['7018,BELLERIVE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7018-BELLERIVE")
df_out = api.search(locations=['7019,CLARENDON VALE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7019-CLARENDON VALE")
df_out = api.search(locations=['7020,CLIFTON BEACH'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7020-CLIFTON BEACH")
df_out = api.search(locations=['7021,LAUDERDALE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7021-LAUDERDALE")
df_out = api.search(locations=['7022,SOUTH ARM'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7022-SOUTH ARM")
df_out = api.search(locations=['7023,OPOSSUM BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7023-OPOSSUM BAY")
df_out = api.search(locations=['7024,CREMORNE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7024-CREMORNE")
df_out = api.search(locations=['7025,DULCOT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7025-DULCOT")
df_out = api.search(locations=['7026,CAMPANIA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7026-CAMPANIA")
df_out = api.search(locations=['7027,COLEBROOK'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7027-COLEBROOK")
df_out = api.search(locations=['7030,APSLEY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7030-APSLEY")
df_out = api.search(locations=['7050,ALBION HEIGHTS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7050-ALBION HEIGHTS")
df_out = api.search(locations=['7052,BLACKMANS BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7052-BLACKMANS BAY")
df_out = api.search(locations=['7053,BONNET HILL'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7053-BONNET HILL")
df_out = api.search(locations=['7054,BARRETTA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7054-BARRETTA")
df_out = api.search(locations=['7055,HUNTINGFIELD'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7055-HUNTINGFIELD")
df_out = api.search(locations=['7109,CATAMARAN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7109-CATAMARAN")
df_out = api.search(locations=['7112,ABELS BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7112-ABELS BAY")
df_out = api.search(locations=['7113,FRANKLIN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7113-FRANKLIN")
df_out = api.search(locations=['7116,BROOKS BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7116-BROOKS BAY")
df_out = api.search(locations=['7117,DOVER'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7117-DOVER")
df_out = api.search(locations=['7119,STONOR'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7119-STONOR")
df_out = api.search(locations=['7120,ANDOVER'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7120-ANDOVER")
df_out = api.search(locations=['7140,BLACK HILLS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7140-BLACK HILLS")
df_out = api.search(locations=['7150,ADVENTURE BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7150-ADVENTURE BAY")
df_out = api.search(locations=['7151,CASEY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7151-CASEY")
df_out = api.search(locations=['7155,KETTERING'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7155-KETTERING")
df_out = api.search(locations=['7162,BIRCHS BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7162-BIRCHS BAY")
df_out = api.search(locations=['7163,FLOWERPOT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7163-FLOWERPOT")
df_out = api.search(locations=['7170,ACTON PARK'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7170-ACTON PARK")
df_out = api.search(locations=['7171,MIDWAY POINT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7171-MIDWAY POINT")
df_out = api.search(locations=['7172,NUGENT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7172-NUGENT")
df_out = api.search(locations=['7173,CARLTON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7173-CARLTON")
df_out = api.search(locations=['7175,BREAM CREEK'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7175-BREAM CREEK")
df_out = api.search(locations=['7176,KELLEVIE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7176-KELLEVIE")
df_out = api.search(locations=['7177,BOOMER BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7177-BOOMER BAY")
df_out = api.search(locations=['7178,MURDUNNA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7178-MURDUNNA")
df_out = api.search(locations=['7179,EAGLEHAWK NECK'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7179-EAGLEHAWK NECK")
df_out = api.search(locations=['7180,TARANNA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7180-TARANNA")
df_out = api.search(locations=['7182,CAPE PILLAR'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7182-CAPE PILLAR")
df_out = api.search(locations=['7184,CAPE RAOUL'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7184-CAPE RAOUL")
df_out = api.search(locations=['7185,PREMAYDENA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7185-PREMAYDENA")
df_out = api.search(locations=['7186,SALTWATER RIVER'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7186-SALTWATER RIVER")
df_out = api.search(locations=['7187,KOONYA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7187-KOONYA")
df_out = api.search(locations=['7190,APSLAWN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7190-APSLAWN")
df_out = api.search(locations=['7209,ROSS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7209-ROSS")
df_out = api.search(locations=['7210,CAMPBELL TOWN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7210-CAMPBELL TOWN")
df_out = api.search(locations=['7211,CLEVELAND'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7211-CLEVELAND")
df_out = api.search(locations=['7212,BEN LOMOND'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7212-BEN LOMOND")
df_out = api.search(locations=['7213,AVOCA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7213-AVOCA")
df_out = api.search(locations=['7214,FINGAL'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7214-FINGAL")
df_out = api.search(locations=['7215,BEAUMARIS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7215-BEAUMARIS")
df_out = api.search(locations=['7216,AKAROA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7216-AKAROA")
df_out = api.search(locations=['7248,ALANVALE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7248-ALANVALE")
df_out = api.search(locations=['7249,GLEN DHU'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7249-GLEN DHU")
df_out = api.search(locations=['7250,BLACKSTONE HEIGHTS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7250-BLACKSTONE HEIGHTS")
df_out = api.search(locations=['7252,BEECHFORD'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7252-BEECHFORD")
df_out = api.search(locations=['7253,BELL BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7253-BELL BAY")
df_out = api.search(locations=['7254,BELLINGHAM'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7254-BELLINGHAM")
df_out = api.search(locations=['7255,BLUE ROCKS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7255-BLUE ROCKS")
df_out = api.search(locations=['7256,BUNGAREE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7256-BUNGAREE")
df_out = api.search(locations=['7257,CAPE BARREN ISLAND'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7257-CAPE BARREN ISLAND")
df_out = api.search(locations=['7258,BREADALBANE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7258-BREADALBANE")
df_out = api.search(locations=['7259,MYRTLE BANK'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7259-MYRTLE BANK")
df_out = api.search(locations=['7260,BLUMONT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7260-BLUMONT")
df_out = api.search(locations=['7261,BRANXHOLM'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7261-BRANXHOLM")
df_out = api.search(locations=['7262,BRIDPORT'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7262-BRIDPORT")
df_out = api.search(locations=['7263,ALBERTON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7263-ALBERTON")
df_out = api.search(locations=['7264,ANSONS BAY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7264-ANSONS BAY")
df_out = api.search(locations=['7265,BANCA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7265-BANCA")
df_out = api.search(locations=['7267,BANGOR'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7267-BANGOR")
df_out = api.search(locations=['7268,LILYDALE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7268-LILYDALE")
df_out = api.search(locations=['7270,BADGER HEAD'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7270-BADGER HEAD")
df_out = api.search(locations=['7275,BLACKWALL'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7275-BLACKWALL")
df_out = api.search(locations=['7276,GRAVELLY BEACH'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7276-GRAVELLY BEACH")
df_out = api.search(locations=['7277,BRIDGENORTH'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7277-BRIDGENORTH")
df_out = api.search(locations=['7290,HADSPEN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7290-HADSPEN")
df_out = api.search(locations=['7291,CARRICK'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7291-CARRICK")
df_out = api.search(locations=['7292,HAGLEY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7292-HAGLEY")
df_out = api.search(locations=['7300,DEVON HILLS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7300-DEVON HILLS")
df_out = api.search(locations=['7301,BISHOPSBOURNE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7301-BISHOPSBOURNE")
df_out = api.search(locations=['7302,BRACKNELL'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7302-BRACKNELL")
df_out = api.search(locations=['7303,BIRRALEE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7303-BIRRALEE")
df_out = api.search(locations=['7304,BRANDUM'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7304-BRANDUM")
df_out = api.search(locations=['7305,MERSEYLEA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7305-MERSEYLEA")
df_out = api.search(locations=['7306,ACACIA HILLS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7306-ACACIA HILLS")
df_out = api.search(locations=['7307,BAKERS BEACH'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7307-BAKERS BEACH")
df_out = api.search(locations=['7310,ABERDEEN'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7310-ABERDEEN")
df_out = api.search(locations=['7315,ABBOTSHAM'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7315-ABBOTSHAM")
df_out = api.search(locations=['7316,CAMENA'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7316-CAMENA")
df_out = api.search(locations=['7320,ACTON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7320-ACTON")
df_out = api.search(locations=['7321,BLACK RIVER'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7321-BLACK RIVER")
df_out = api.search(locations=['7322,SOMERSET'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7322-SOMERSET")
df_out = api.search(locations=['7325,CALDER'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7325-CALDER")
df_out = api.search(locations=['7330,ALCOMIE'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7330-ALCOMIE")
df_out = api.search(locations=['7331,STANLEY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7331-STANLEY")
df_out = api.search(locations=['7466,GORMANSTON'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7466-GORMANSTON")
df_out = api.search(locations=['7467,LAKE MARGARET'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7467-LAKE MARGARET")
df_out = api.search(locations=['7468,MACQUARIE HEADS'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7468-MACQUARIE HEADS")
df_out = api.search(locations=['7469,GRANVILLE HARBOUR'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7469-GRANVILLE HARBOUR")
df_out = api.search(locations=['7470,ROSEBERY'] )
df_out = pd.DataFrame(df_out)
DF_ALL = DF_ALL.append(df_out, ignore_index=True) 
print("finished all:7470-ROSEBERY")

DF_ALL1 = DF_ALL.reset_index()
DF_ALL2 = DF_ALL1.drop(columns=['id','index','images','images_floorplans','listers'])
DF_ALL_UNIQ = DF_ALL2.drop_duplicates()

DF_ALL.to_pickle("/Users/joezhou/Downloads/Realestate/R_ALL_TAS.pkl")
