==> 1_0 Sample of 1.ipynb <==
pip install numpy
pip install pandas
pip install matplotlib


# pip install yfinance --upgrade --no-cache-dir
# pip install yfinance


import yfinance as yf
dat = yf.Ticker("CBA.AX")

dat = yf.Ticker("CBA.AX")
dat.info
dat.calendar
dat.analyst_price_targets
dat.quarterly_income_stmt
dat.history(period='1mo')
# dat.option_chain(dat.options[0]).calls
