==> 2_1 Feature - Info.ipynb <==
# The aim of this module is to keep non time sensitive variables

# File name convention
# R_ = raw extract 
# E_ = Entry files into enrichment layer
# C_ = enriched layer with calculation
# I = insights layer, designed for model baselines
# V_ = validation

import pandas as pd

# DF_YFIN_INFO = pd.read_csv (r'/Users/joezhou/Downloads/R_ALL_INFO.csv',sep='|') 

# DF_YFIN_INFO1 = pd.read_csv (r'/Users/joezhou/Downloads/Investment/1 data/R_ALL_INFO_1.csv',sep='|') 
# DF_YFIN_INFO2 = pd.read_csv (r'/Users/joezhou/Downloads/Investment/1 data/R_ALL_INFO_2.csv',sep='|') 
# DF_YFIN_INFO3 = pd.read_csv (r'/Users/joezhou/Downloads/Investment/1 data/R_ALL_INFO_3.csv',sep='|') 
# DF_YFIN_INFO4 = pd.read_csv (r'/Users/joezhou/Downloads/Investment/1 data/R_ALL_INFO_4.csv',sep='|') 

# DF_YFIN_INFO1 = pd.read_pickle ('/Users/joezhou/Downloads/R_ALL_INFO_1.pkl') 
# DF_YFIN_INFO2 = pd.read_pickle ('/Users/joezhou/Downloads/R_ALL_INFO_2.pkl') 
# DF_YFIN_INFO3 = pd.read_pickle ('/Users/joezhou/Downloads/R_ALL_INFO_3.pkl') 
# DF_YFIN_INFO4 = pd.read_pickle ('/Users/joezhou/Downloads/R_ALL_INFO_4.pkl') 

DF_YFIN_INFO = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_INFO.pkl') 
DF_YFIN_INFO.ignore_index=True


# DF_YFIN_INFO = pd.concat([DF_YFIN_INFO1, DF_YFIN_INFO2, DF_YFIN_INFO3, DF_YFIN_INFO4], ignore_index=True)

# del DF_YFIN_INFO1, DF_YFIN_INFO2, DF_YFIN_INFO3, DF_YFIN_INFO4

# extract list of columns for selection
DF_YFIN_INFO_COLS = list(DF_YFIN_INFO.columns)

# find the number of missing row
DF_YFIN_INFO_NAN = DF_YFIN_INFO.isna().sum()


from datetime import datetime as dt
dte = dt.now().strftime('%Y%m%d')
fname = "/Users/joezhou/Downloads/NAN_COUNT_R_YFIN_INFO_{}.xlsx".format(dte)
DF_YFIN_INFO_NAN.to_excel(fname)




<!-- Columns with majority of 99%+ of values -->
['algorithm'],['annualHoldingsTurnover'],['annualReportExpenseRatio'],['category'],['circulatingSupply'],['dateShortInterest'],['expireDate'],['fromCurrency'],['lastCapGain'],['lastMarket'],['maxSupply'],['morningStarOverallRating'],['morningStarRiskRating'],['revenueQuarterlyGrowth'],['sharesPercentSharesOut'],['sharesShort'],['sharesShortPreviousMonthDate'],['sharesShortPriorMonth'],['shortPercentOfFloat'],['shortRatio'],['startDate'],['strikePrice'],['toCurrency'],['volume24Hr'],['volumeAllCurrencies'],['ytdReturn'],['headSymbol'],['underlyingExchangeSymbol'],['underlyingSymbol'],['postMarketChange'],['postMarketPrice'],['preMarketChange'],['preMarketPrice'],['regularMarketChange'],['averageDailyVolume3Month'],['beta3Year'],['fiveYearAverageReturn'],['fundFamily'],['fundInceptionDate'],['legalType'],['navPrice'],['threeYearAverageReturn'],['totalAssets'],['yield'],['openInterest'],['lastDividendDate'],['lastDividendValue'],['exchangeDataDelayedBy'],['exchangeName'],['marketState'],['quoteSourceName'],['regularMarketSource'],['regularMarketTime'],['52WeekChange'],['SandP52WeekChange'],['impliedSharesOutstanding'],

<!-- Columns to remove due to lack of information value -->
<!-- location -->
['address1'],['companyOfficers'],['address2'],
<!-- Misc -->
['currency'],['exchange'],['exchangeTimezoneName'],['exchangeTimezoneShortName'],['gmtOffSetMilliseconds'],['market'],['maxAge'],['tradeable'],
<!-- Price -->
['ask'],['bid'],['dayHigh'],['dayLow'],['fiftyDayAverage'],['fiftyTwoWeekHigh'],['fiftyTwoWeekLow'],['open'],['previousClose'],['regularMarketDayHigh'],['regularMarketDayLow'],['regularMarketOpen'],['regularMarketPreviousClose'],['regularMarketPrice'],['regularMarketVolume'],['volume'],['currentPrice'],['askSize'],['bidSize'],
<!-- Reference -->
['longBusinessSummary'],['logo_url'],['website'],['financialCurrency'],['phone'],['fax'],
<!-- Timing -->
['mostRecentQuarter'],['lastFiscalYearEnd'],['nextFiscalYearEnd'],['lastSplitDate'],['lastSplitFactor'],['exDividendDate'],

<!-- Reviewed on 20220724, 169 variables were in the information extract -->

DF_YFIN_INFO_SUBSET = DF_YFIN_INFO[[

# Identifiers
'symbol','messageBoardId','quoteType','shortName','longName','uuid',
# Location
'city','country','zip','state',
# Industry and size
'industry','sector','fullTimeEmployees',

#holder details
'heldPercentInsiders','heldPercentInstitutions',

# Dividend related
'trailingAnnualDividendRate','trailingAnnualDividendYield','dividendRate','dividendYield','fiveYearAvgDividendYield',

# variables to investigate
# 'isEsgPopulated','priceHint','recommendationKey','numberOfAnalystOpinions',

# Price movement and targets
'averageDailyVolume10Day','averageVolume','averageVolume10days',
'twoHundredDayAverage','targetHighPrice','targetLowPrice','targetMeanPrice','targetMedianPrice','recommendationMean',

# Financial metrics
# forecast and market influence
'forwardEps','forwardPE','beta',
# Operations
'earningsGrowth','earningsQuarterlyGrowth','ebitda','ebitdaMargins','freeCashflow','grossMargins','grossProfits','netIncomeToCommon','operatingCashflow','operatingMargins','payoutRatio','profitMargins','revenueGrowth','totalRevenue',
# Portfolio
'bookValue','currentRatio','debtToEquity','enterpriseValue','quickRatio','returnOnAssets','returnOnEquity','totalCash','totalDebt',
# price valuation
'pegRatio','priceToBook','trailingPE','enterpriseToEbitda','enterpriseToRevenue','priceToSalesTrailing12Months',
# unit value
'revenuePerShare','totalCashPerShare','trailingEps',
# volume
'floatShares','marketCap','sharesOutstanding',

]]

DF_YFIN_INFO_SUBSET = DF_YFIN_INFO_SUBSET.rename(columns={"symbol": "TickName"})




DF_YFIN_INFO_SUBSET.to_pickle("/Users/joezhou/Downloads/Investment/1 data/E_ALL_INFO.pkl")
