==> 2_4 Feature - Financials.ipynb <==
# The aim of this module is to calculate features using financial information

# File name convention
# R_ = raw extract 
# E_ = Entry files into enrichment layer
# C_ = enriched layer with calculation
# I = insights layer, designed for model baselines
# V_ = validation

import pandas as pd
import numpy as np

# FIN = pd.read_csv (r'/Users/joezhou/Downloads/R_ALL_FIN.csv',sep='|') 
FIN1 = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_FIN_1.pkl') 
FIN2 = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_FIN_2.pkl') 
FIN3 = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_FIN_3.pkl') 
FIN4 = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_FIN_4.pkl') 

FIN = pd.concat([FIN1, FIN2, FIN3, FIN4], ignore_index=True)
del FIN1, FIN2, FIN3, FIN4

#Create a sample for code testing purposes
# FIN = DF[DF['TickName'].isin(['CBA.AX','ANZ.AX','APT.AX','Z1P.AX'])]

#create a baselist for adding on the feature engineered variables
TICKER_LIST = FIN['TickName'].drop_duplicates().reset_index()


# add suffix to the metric name before transforming into columns
# df = FIN
# df['color'] = np.where(df['Financial Data Type']== "Income Statement", 'F_INC_','')
# df['color'] = np.where(df['Financial Data Type']== "Balance Sheet", 'F_BS_','')
# df['color'] = np.where(df['Financial Data Type']== "Cashflow", 'F_CF_'+df['Metric'],'')

# df['Metric_N'] = np.where(df['Financial Data Type']== "Income Statement", 'F_INC_'+df['Metric'],np.where(df['Financial Data Type']== "Balance Sheet", 'F_BS_'+df['Metric'],'F_CF_'+df['Metric']))

# del df['Metric']


# Transpose to ticker and date grain
FIN['Value_num']=FIN['Value'].fillna(0).astype(int)
FIN_METRICS = FIN.pivot_table(index=['TickName', 'Date'], columns='Metric', values='Value_num').reset_index()

FIN_METRICS_LAST = FIN_METRICS.groupby("TickName").last()


# add some mmentum indicators
# contine revenue growth
# tax rate
# common ratios, asset to liab etc, cashflow available vs using

# Export for the next step
FIN_METRICS_LAST.to_pickle("/Users/joezhou/Downloads/Investment/1 data/E_ALL_FIN.pkl")


from datetime import datetime as dt
dte = dt.now().strftime('%Y%m%d')
fname = "/Users/joezhou/Downloads/Investment/TEST_FIN_Last{}.xlsx".format(dte)
FIN_METRICS_LAST.to_excel(fname)
