==> 3_x Triggers to Investigate.ipynb <==
The objective of this module is to identify triggers for further research

import pandas as pd
import numpy as np


# The aim of this module is to calculate features using financial information

# File name convention
# R_ = raw extract 
# E_ = Entry files into enrichment layer
# C_ = enriched layer with calculation
# I = insights layer, designed for model baselines
# V_ = validation

R_INFO = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_INFO.pkl') 
R_DIV = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_Div.pkl') 
R_PRICE = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_Prices.pkl') 

FIN1 = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_FIN_1.pkl') 
FIN2 = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_FIN_2.pkl') 
FIN3 = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_FIN_3.pkl') 
FIN4 = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_FIN_4.pkl') 

R_FIN = pd.concat([FIN1, FIN2, FIN3, FIN4], ignore_index=True)
del FIN1, FIN2, FIN3, FIN4



I_TIC = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/E_TICK_INPUTS.pkl') 
I_INFO = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/E_ALL_INFO.pkl') 
I_DIV = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/E_ALL_Div.pkl') 
I_PRICE = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/E_ALL_Prices.pkl') 
I_FIN = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/E_ALL_FIN.pkl') 

I_FLAT = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/C_TICKER_FLAT.pkl') 



from datetime import datetime as dt
dte = dt.now().strftime('%Y%m%d')
fname = "/Users/joezhou/Downloads/I_FLAT{}.xlsx".format(dte)
I_FLAT.to_excel(fname)
