==> 3_2 Calc discount rate.ipynb <==
# The aim of this module is to calculate the discout rate using a weighted cost of capital approach

# File name convention
# R_ = raw extract 
# E_ = Entry files into enrichment layer
# C_ = enriched layer with calculation
# I = insights layer, designed for model baselines
# V_ = validation


import pandas as pd
import numpy as np
pd.options.mode.chained_assignment = None  # default='warn'

I_FLAT = pd.read_pickle ('/Users/joezhou/Downloads/C_TICKER_FLAT.pkl') 
C_BG_MODEL = pd.read_pickle ('/Users/joezhou/Downloads/C_BEN_G_MODEL.pkl') 




#Create a sample for code testing purposes
FIN_S = FIN[FIN['TickName'].isin(['WES.AX'])]
FIN_S = FIN_S.loc[(FIN_S['Date'] == '2020-06-30')]


# Caclulate WACC for Westfarmers

FIN_S_T = FIN_S.pivot_table('Value', ['TickName', 'Date'], 'Metric')

# FIN_S_T.to_excel('/Users/joezhou/Downloads/test_file.xlsx', engine='xlsxwriter')  

FIN_S_T['C_TAX_R'] = FIN_S_T['Income Tax Expense']/ FIN_S_T['Income Before Tax']
FIN_S_T['C_TOT_DEBT'] = FIN_S_T['Short Long Term Debt']+FIN_S_T['Long Term Debt']
FIN_S_T['C_TOT_CPTL'] = FIN_S_T['C_TOT_DEBT'] + FIN_S_T['Total Stockholder Equity']

FIN_S_T['C_TOT_DEBT_PERC'] = FIN_S_T['C_TOT_DEBT'] / FIN_S_T['C_TOT_CPTL']
FIN_S_T['C_TOT_EQTY_PERC'] = 1 - FIN_S_T['C_TOT_DEBT_PERC']

FIN_S_T['C_COST_DEBT'] = FIN_S_T['Long Term Debt'] / FIN_S_T['Total Assets']
#need to work out cost of equity using westfarmers excel sheet
FIN_S_T['C_COST_EQTY'] = 0.0502

FIN_S_T['C_WACC'] = (FIN_S_T['C_TOT_DEBT_PERC']* FIN_S_T['C_COST_DEBT']) + (FIN_S_T['C_TOT_EQTY_PERC']*FIN_S_T['C_COST_EQTY'])*(1-FIN_S_T['C_TAX_R'])


# extract variables and formatting
FIN_S2 = FIN_S[FIN_S['Metric'].isin(['Short Long Term Debt','Total Stockholder Equity'])]
# the above used to calculate debt and equity rate

# risk free rate, use the 10 year bond rate
RF = .01

# BETA = use from the info section

# Gross yield
average_return = 5.63




# BG_MODEL.to_pickle("/Users/joezhou/Downloads/C_CALCS.pkl")
# BG_MODEL.to_excel('/Users/joezhou/Downloads/C_CALCS.xlsx', engine='xlsxwriter')
