==> 4_0 All combined.ipynb <==
# The aim of this module is to combine all features

# File name convention
# R_ = raw extract 
# E_ = Entry files into enrichment layer
# C_ = enriched layer with calculation
# I = insights layer, designed for model baselines
# V_ = validation

import pandas as pd
import numpy as np
pd.options.mode.chained_assignment = None  # default='warn'

C_FLAT = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/C_TICKER_FLAT.pkl') 

C_VALUE = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/C_VALUATION.pkl') 

C_BG_MOD = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/C_BEN_G_MODEL.pkl') 
C_BG_MOD = C_BG_MOD[['TickName','CALC_BG_CONS_BG_SCORE','CALC_BG_ENT_BG_SCORE']]


DF_VAL = C_VALUE.merge(C_BG_MOD,how='inner',on='TickName').merge(C_FLAT,how='inner',on='TickName')
DF_VAL['CALC_VAL_DIF']=DF_VAL['CALC_VAL_NTA_DIV']-DF_VAL['Last_Close_price']
DF_VAL['CALC_VAL_DIF_PERC']=DF_VAL['CALC_VAL_DIF']/DF_VAL['Last_Close_price']

DF_VAL['CALC_DIV_PAYOUT_RATE'] = DF_VAL['CALC_DIV_3YR_AMT']/3 / DF_VAL['Last_Close_price']



DF_VAL.to_pickle("/Users/joezhou/Downloads/Investment/1 data/I_MASTER_FLAT.pkl")



from datetime import datetime as dt
dte = dt.now().strftime('%Y%m%d')
fname = "/Users/joezhou/Downloads/Investment/TEST_FLAT_MASTER{}.xlsx".format(dte)
DF_VAL.to_excel(fname)

fname = "/Users/joezhou/Downloads/Investment/MASTER_DATA_LATEST.xlsx".format(dte)
DF_VAL.to_excel(fname)



from datetime import datetime as dt
dte = dt.now().strftime('%Y%m%d')
fname = "/Users/joezhou/Downloads/Investment/TEST_C_FLAT{}.xlsx".format(dte)
C_FLAT.to_excel(fname)


TEST_DIV = pd.read_pickle ('/Users/joezhou/Downloads/Investment/1 data/R_ALL_DIV.pkl')
