==> 4_1 Analysis - clustering.ipynb <==
# Create clusters from the reduced set
from sklearn.preprocessing import MinMaxScaler
from sklearn.cluster import KMeans

#Split into category vs continous features
categorical_features = DF_YFIN_INFO_BASE[['sector','state']]
continuous_features = DF_YFIN_INFO_BASE[['dividendRate',
'heldPercentInstitutions',
'beta',
'profitMargins',
'trailingPE']]

data = DF_YFIN_INFO_BASE

# convert categories to numeric
for col in categorical_features:
    dummies = pd.get_dummies(data[col], prefix=col)
    data = pd.concat([data, dummies], axis=1)
    data.drop(col, axis=1, inplace=True)

# Standardise data
import numpy as np

data.replace([np.inf, -np.inf], np.nan, inplace=True)


# standardise continuous variables
mms = MinMaxScaler()
mms.fit(data)
data_transformed = mms.transform(data)

# fill in NaN data points for the array
data_transformed[np.isnan(data_transformed)] = 0



Sum_of_squared_distances = []
K = range(1,15)
for k in K:
    km = KMeans(n_clusters=k)
    km = km.fit(data_transformed)
    Sum_of_squared_distances.append(km.inertia_)

plt.plot(K, Sum_of_squared_distances, 'bx-')
plt.xlabel('k')
plt.ylabel('Sum_of_squared_distances')
plt.title('Elbow Method For Optimal k')
plt.show()


df = data_transformed
  
kmeans = KMeans(n_clusters=4).fit(df)
centroids = kmeans.cluster_centers_
print(centroids)

plt.scatter(df['x'], df['y'], c= kmeans.labels_.astype(float), s=50, alpha=0.5)
plt.scatter(centroids[:, 0], centroids[:, 1], c='red', s=50)
plt.show()

# Find Principal components

x = continuous_features.dropna()

from sklearn.decomposition import PCA
pca = PCA(n_components=2)
principalComponents = pca.fit_transform(x)
principalDf = pd.DataFrame(data = principalComponents
             , columns = ['principal component 1', 'principal component 2'])



principalDf.head()

principalDf.plot('principal component 1','principal component 2',kind = 'scatter')


fig = plt.figure(figsize = (8,8))
ax = fig.add_subplot(1,1,1) 
ax.set_xlabel('Principal Component 1', fontsize = 15)
ax.set_ylabel('Principal Component 2', fontsize = 15)
ax.set_title('2 component PCA', fontsize = 20)
targets = ['Iris-setosa', 'Iris-versicolor', 'Iris-virginica']
colors = ['r', 'g', 'b']
for target, color in zip(targets,colors):
    indicesToKeep = finalDf['target'] == target
    ax.scatter(finalDf.loc[indicesToKeep, 'principal component 1']
               , finalDf.loc[indicesToKeep, 'principal component 2']
               , c = color
               , s = 50)
ax.legend(targets)
ax.grid()

# Answer the following Questions
1. How does it compare to the market
    Standardise movements
    Rank the volatility
