==> 4_1 Analysis - Historical prices.ipynb <==
import pandas as pd

# Import data from excel, switch to sqlite later on
# INF_DF = pd.read_excel (r'/Users/JZ/Downloads/TEST_INFO.xlsx')
# FINANCIALS_DF = pd.read_excel (r'/Users/JZ/Downloads/TEST_FIN.xlsx')
HIST_PRICE_DF = pd.read_csv (r'/Users/JZ/Downloads/ALL_Prices.csv', sep='|')


DF_DUP_CHECK = HIST_PRICE_DF.groupby('TickName')['TickName'].nunique()

#Find Correlation between all listed companies
HIST_PRICE_DF1 = HIST_PRICE_DF
#flip data columns
HIST_PRICE_DF1['Daily_Open_Close_Chg_Rate']=(HIST_PRICE_DF1['Close']-HIST_PRICE_DF1['Open'])/HIST_PRICE_DF1['Open']
HIST_PRICE_DF1['Daily_High_Low_Chg_Rate']=(HIST_PRICE_DF1['High']-HIST_PRICE_DF1['Low'])/HIST_PRICE_DF1['Open']
#HIST_PRICE_DF1.pivot( columns='TickName')['Daily_Open_Close_Chg_Rate']
