==> 4_2 Dividend Forecast.ipynb <==
import numpy as np
import pandas as  pd
import matplotlib.pyplot as plt
import seaborn as sns

#!pip install yfinance --upgrade --no-cache-dir
#!pip uninstall yfinance
#!pip install git+https://github.com/rodrigobercini/yfinance.git
import yfinance as yf
import yahoofinancials as yfin

tick = yf.Ticker("CBA.AX")

DIV = tick.dividends
DIV_PA = DIV.resample('Y').sum()
# minimum of last 4 dividends excluding the current calendar year
# current calendar year is represented by -1
DIV_L4_AVG = DIV_PA.iloc[[-5,-4,-3,-2]].sum()/4
DIV_L4_STD = DIV_PA.iloc[[-5,-4,-3,-2]].std()
DIVIDEND_BASE = DIV_L4_AVG - DIV_L4_STD
DIVIDEND_BASE

#find last 5 years of dividends
DIV2 = DIV
DIV3 = DIV2.resample('Y').sum()
DIV4= DIV3.iloc[[-5,-4,-3,-2]]
DIV4

DIV2 = DIV

DIV3 = DIV2.resample('Y').sum()
mean = DIV3[-6:-1].mean()
std = DIV[-6:-1].std()
#DIV3[:-1].plot(title="yearly CBA Dividend over time")
print(mean,std)
