==> 4_3 Sustainability Impact.ipynb <==
#Download sustainability data and find correlation between the score and stock prices

import numpy as np
import pandas as  pd
import matplotlib.pyplot as plt
import seaborn as sns

#!pip install yfinance --upgrade --no-cache-dir
#!pip uninstall yfinance
#!pip install git+https://github.com/rodrigobercini/yfinance.git
import yfinance as yf
import yahoofinancials as yfin

import pandas as pd
ASX_data = pd.read_csv('https://www.asx.com.au/asx/research/ASXListedCompanies.csv',skiprows=1)
# format to align with yahoo finance ticker
ASX_data["Ticker"] = ASX_data['ASX code']+".AX"
ASX_TICKER_LIST = list(ASX_data['Ticker'])

import time
start_time = time.time()

company_df = pd.DataFrame()

def Info_Extract(TickName):
    
    global company_df
    
    try:
        tick = yf.Ticker(TickName)
        tick_sustain = tick.sustainability.transpose()
        tick_sustain['TICKER'] = TickName
    
        company_df = company_df.append(tick_sustain, ignore_index=True) 
         
    except Exception:
        print("Error with",TickName)
        pass

    return()

for company in ASX_TICKER_LIST:
    Info_Extract(company)

#company_df.to_excel("/Users/JZ/Downloads/ALL_ORDS_500_SUSTAINABILITY.xlsx", sheet_name='Sheet_name_1')

print("--- %s seconds ---" % (time.time() - start_time))
